Move `getChildEntity` and `buildEntityPath` to new hdf5-utils.ts
