+++ packages/app/src/metadata-viewer/MetadataViewer.tsx
@@ -1,6 +1,6 @@
 import { isAbsolutePath, isDataset } from '@h5web/shared/guards';
 import { EntityKind } from '@h5web/shared/hdf5-models';
-import { buildEntityPath } from '@h5web/shared/utils';
+import { buildEntityPath } from '@h5web/shared/hdf5-utils';
 import { capitalize } from 'lodash';
 import { memo, Suspense } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
