+++ packages/app/src/providers/hsds/hsds-api.ts
@@ -16,7 +16,7 @@ import type {
   Value,
 } from '@h5web/shared/hdf5-models';
 import { EntityKind } from '@h5web/shared/hdf5-models';
-import { buildEntityPath, getChildEntity } from '@h5web/shared/utils';
+import { buildEntityPath, getChildEntity } from '@h5web/shared/hdf5-utils';
 
 import { DataProviderApi } from '../api';
 import type { ExportFormat, ExportURL, ValuesStoreParams } from '../models';
