+++ packages/shared/src/hdf5-utils.ts
@@ -1,0 +1,16 @@
+import type { ChildEntity, GroupWithChildren } from './hdf5-models';
+
+export function getChildEntity(
+  group: GroupWithChildren,
+  entityName: string,
+): ChildEntity | undefined {
+  return group.children.find((child) => child.name === entityName);
+}
+
+export function buildEntityPath(
+  parentPath: string,
+  entityNameOrRelativePath: string,
+): string {
+  const prefix = parentPath === '/' ? '' : parentPath;
+  return `${prefix}/${entityNameOrRelativePath}`;
+}
